requirements.txt
fastapi
uvicorn
openai
pymongo
pydantic
python-dotenv
sentence-transformers
python-jose[cryptography]
pydantic[email]